refactor: Move MaybeTooltip to the components folder
