+++ app/components/form.tsx
@@ -41,6 +41,7 @@ import React, {
 } from "react";
 
 import { useBrowseContext } from "@/browser/context";
+import { MaybeTooltip } from "@/components/maybe-tooltip";
 import { BANNER_MARGIN_TOP } from "@/components/presence";
 import VisuallyHidden from "@/components/visually-hidden";
 import {
@@ -52,7 +53,6 @@ import {
 import { Icon } from "@/icons";
 import SvgIcExclamation from "@/icons/components/IcExclamation";
 import { useLocale } from "@/locales/use-locale";
-import { MaybeTooltip } from "@/utils/maybe-tooltip";
 import { valueComparator } from "@/utils/sorting-values";
 
 export const Label = ({
